rootfs/requirements.txt
PyYAML==6.0.2
gunicorn==23.0.0
openbrokerapi==4.7.1
requests==2.32.2
celery==5.4.0
redis==5.2.0
jsonschema==4.23.0
